Implement POC version of module federation
