+++ package.json
@@ -3,7 +3,7 @@
   "browser": {
     "./client/web_client.jsx": "./client/browser_web_client.jsx"
   },
-  "version": "7.0.0",
+  "version": "7.3.0-0",
   "private": true,
   "engines": {
     "npm": ">=7.0.0",
